Removed 'simplify' option, we should stay consistent.
For the sake of consistency and to be future-proof, we don’t allow to
choose between simplified dimension names or not. Flat dimensions are
called just by their dimension names and mode detailed dims have format
dim.attribute. In the future ‘dim` withoiut `attribute` would refer to
the dimension key (once hierarchies are removed).