Fix import regression which disallow to `slicer model validate`.
`cubes.metadata.*` is added to `cubes.metadata` because the validate_model
was previously a function part of it (ascendance compatibility).

Then, it impacts necessarily the new import added to
`cubes.slicer.commands`.

In the function `cubes.slicer.commands`, the packages `cubes` and
`cubes.providers` are no longer available, the link to the owned
functions are so rebuild.